+++ src/components/Main/CardDivide.js
@@ -1,0 +1,17 @@
+import Card from "../UI/Card";
+
+const CardDivide = ({ folderCards }) => {
+  return (
+    <>
+      {folderCards?.map((folderCard) => {
+        return (
+          <div key={folderCard.id}>
+            <Card item={folderCard} />
+          </div>
+        );
+      })}
+    </>
+  );
+};
+
+export default CardDivide;
